Set up initial script
The first draft of this script iterates through the files of a local directory
containing input photos/screenshots. The location of each photo is extracted via
GPT-4 Vision and used to retrieve the corresponding Google Maps data.

Testing:
Ran with test folder containing screenshots of Instagram stories that have
location tags and confirmed that returned data is as expected.