Don't add duplicate locations to Notion database
If we've already saved a location to the database, we shouldn't add it again.

Also tweaked the prompt to extract supplemental information that we can attach
to the Google Maps query to improve our chances of retrieving the right place.

Testing:
Ran on Instagram screenshots (not stories). Correctly identified places with
ambiguous names.