Save locations to Google Maps "Want to go" list
Save the extracted locations to user's "Want to go" list in Google Maps rather
than a Notion database. They'll be asked to log in to their Google account, and
the script will take over from there to save all the places individually.

Testing:
Ran script and confirmed that locations were saved to "Want to go", or skipped
if they'd been previously saved.