Adjust prompt to account for people/companies
With this change, we'll now recognize different types of subjects in the
screenshots, namely people and companies in addition to locations. In a
future change, these people/companies will be added to a Notion
database. The prompt has been rewritten to be more clear and easy to
understand.

Also account for different file types (previously just assumed all input
images were jpegs).

Testing:
Added screenshots of Twitter profiles of people/companies and confirmed
that prompt output looks as expected.