Include note in Serp query when retrieving links
To increase our chances of grabbing the correct corresponding link for
each entity, we'll include additional context from the note when calling
Serp API for the Google Search results rather than just relying on the
name.

Testing:
Confirmed that one of the companies which was previously getting linked
to the wrong place is now attributed to the right page.